// src/types/Hero.ts
export default interface Hero {
    name: string;
    power: number;
    agility: number;
    intelligence: number;
    charisma: number;
    takenDamage: number;
    attack: number;
    stealth: number;
    archery: number;
    learning: number;
    survivability: number;
    medicine: number;
    intimidation: number;
    insight: number;
    appearance: number;
    manipulation: number;
}